requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.8.0
django-environ == 0.9.0
django-modeltranslation==0.19.17
django-parler==2.3
django-rosetta==0.10.2
django-seed==0.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==37.8.0
idna==3.10
polib==1.2.0
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
toposort==1.10
tzdata==2025.2
urllib3==2.5.0
